added csv write

+++ pyproc-server.py
@@ -1,17 +1,22 @@
 import os
 from datetime import datetime
+import csv
+import sys
 
 ansap = 0
 metap = 0
 animatorp = 0
 tmp = datetime.now()
 now = tmp.strftime("%m%d%Y-%H%M")
+nowh = tmp.strftime("%m.%d.%Y-%H")
+
 
 #smaze info o starych informacich
 def delfiles():
 	if len(os.listdir('/www/')) > 0:
 		for files in os.listdir('/www/'):
-			os.remove('/www/{}'.format(files))
+			if not files.endswith('data.csv'):
+				os.remove('/www/{}'.format(files))
 
 #fce ktera vytahne procesy ze vsech souboru a zapise je do jednoho
 def getprocs():
@@ -46,11 +51,27 @@ def writeprocs():
 		file.write('meta - {}\n'.format(metap))
 		file.write('animator - {}\n'.format(animatorp))
 
+def writetocsv():
+	with open('/www/data.csv', 'a') as csv_file:
+		writecsv = csv.writer(csv_file)
+		writecsv.writerow([nowh, ansap, metap, animatorp])
+	csv_file.close()
+
+def checkcsv():
+	with open('/www/data.csv', 'r') as csv_check:
+		next(csv_check)
+		check = csv.reader(csv_check, delimiter=',')
+		for line in check:
+			if nowh not in line[0]:
+				writetocsv()
+
+
 def main():
 	getprocs()
 	countprocs()
 	delfiles()
 	writeprocs()
+	checkcsv()
 
 
 if __name__ == "__main__":
